+++ CHANGELOG.md
@@ -45,6 +45,8 @@ Version 2.99.x and version 3.x will be available with a [pro license](http://www
 
 **Next release**
 
+* Updated Slovenian translation
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.99.26 BETA**
@@ -154,6 +156,7 @@ XPrivacy 2
 * Updated Dutch translation
 * Updated German translation
 * Updated Japanese translation
+* Updated Slovenian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)